requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
gunicorn>=21.2.0

# Validation and Settings
pydantic>=2.4.2
pydantic-settings>=2.0.3
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9  # PostgreSQL adapter
asyncpg>=0.29.0  # Async PostgreSQL adapter
alembic>=1.12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4  # Specific version for compatibility
bcrypt==4.0.1  # Added specific version for compatibility with passlib
python-multipart==0.0.12

# Azure
azure-identity==1.19.0
azure-storage-blob==12.23.1

# Utilities
python-dotenv==1.0.1
qrcode[pil]==7.4.2
pillow==10.4.0
pandas==2.2.3
